Refactoring of MatchedSequencesEncoder, more caching
